// Code/UICode/src/app/Interfaces/alertDetails.ts
export class AlertDetails {
    Beacon_ID: any;
    blockTimestamp: any;
    Gateway_ID: any;
    Shipment_ID: any;
    Pallet_ID: any;
    Alert_Start_Time: any;
    Alert_End_Time: any;
    Alert_Loc: any;
    Alert_Type: any;
    Alert_Value: any;
    Alert_Lat: any;
    Alert_Long: any;
    // Last_GPS_Time: any;
    // Alert_System_Time: any;
    Object_ID: any;
    Object_Type: any;
}